requirements.txt
django
djangorestframework
djangorestframework-simplejwt
djongo==1.3.6
pymongo==3.12.3

